crypto/bbb/solver/requirements.txt
pwntools == 4.8.0
pycryptodome == 3.15.0